Adds a myth2kodi version number.
Will update this in the commit following a release/deployment.
So the commit prior to the next version bump will be the "official"
Version 0.1.0. Intend to change the minor number for fixes to
existing functionality, middle number for new features implemented,
arriving at 1.0.0 once all currently planned features are implemented.